+++ src/components/Button/Button.tsx
@@ -8,30 +8,32 @@ import { ButtonBehavior } from '../../lib/accessibility'
 import { Accessibility } from '../../lib/accessibility/interfaces'
 import { ComponentVariablesInput, ComponentPartStyle } from '../../../types/theme'
 import {
-  Extendable,
-  ItemShorthand,
-  ReactChildren,
   ComponentEventHandler,
+  Extendable,
+  ReactChildren,
+  ShorthandRenderFunction,
+  ShorthandValue,
 } from '../../../types/utils'
 import ButtonGroup from './ButtonGroup'
 import isFromKeyboard from '../../lib/isFromKeyboard'
 
 export interface IButtonProps {
   as?: any
+  accessibility?: Accessibility
   children?: ReactChildren
   circular?: boolean
   className?: string
   disabled?: boolean
   content?: React.ReactNode
   fluid?: boolean
-  icon?: ItemShorthand
+  icon?: ShorthandValue
   iconOnly?: boolean
   iconPosition?: 'before' | 'after'
   onClick?: ComponentEventHandler<IButtonProps>
   onFocus?: ComponentEventHandler<IButtonProps>
+  renderIcon?: ShorthandRenderFunction
   text?: boolean
   type?: 'primary' | 'secondary'
-  accessibility?: Accessibility
   styles?: ComponentPartStyle
   variables?: ComponentVariablesInput
 }
@@ -111,6 +113,15 @@ class Button extends UIComponent<Extendable<IButtonProps>, IButtonState> {
     /** Accessibility behavior if overridden by the user. */
     accessibility: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
 
+    /**
+     * A custom render function the icon slot.
+     *
+     * @param {React.ReactType} Component - The computed component for this slot.
+     * @param {object} props - The computed props for this slot.
+     * @param {ReactNode|ReactNodeArray} children - The computed children for this slot.
+     */
+    renderIcon: PropTypes.func,
+
     /** Additional CSS styles to apply to the component instance.  */
     styles: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
 
@@ -156,13 +167,14 @@ class Button extends UIComponent<Extendable<IButtonProps>, IButtonState> {
   }
 
   public renderIcon = (variables, styles) => {
-    const { icon, iconPosition, content } = this.props
+    const { icon, iconPosition, content, renderIcon } = this.props
 
     return Icon.create(icon, {
       defaultProps: {
         styles: styles.icon,
         xSpacing: !content ? 'none' : iconPosition === 'after' ? 'before' : 'after',
         variables: variables.icon,
+        render: renderIcon,
       },
     })
   }